# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zoom Meeting Automation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UpdateProfileInput {
  name?: string;
  companyName?: string;
  jobTitle?: string;
  phone?: string;
}

export type MeetingMeetingType = typeof MeetingMeetingType[keyof typeof MeetingMeetingType];


export const MeetingMeetingType = {
  client: 'client',
  internal: 'internal',
  personal: 'personal',
} as const;

export type MeetingStatus = typeof MeetingStatus[keyof typeof MeetingStatus];


export const MeetingStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Meeting {
  id: string;
  userId: string;
  meetingType: MeetingMeetingType;
  clientName: string;
  clientEmail: string;
  cc: string[];
  bcc: string[];
  meetingTitle: string;
  /** @nullable */
  description?: string | null;
  meetingDate: string;
  meetingTime: string;
  duration: number;
  timezone: string;
  /** @nullable */
  organizerName?: string | null;
  /** @nullable */
  organizerEmail?: string | null;
  /** @nullable */
  organizerCompany?: string | null;
  /** @nullable */
  organizerTitle?: string | null;
  /** @nullable */
  zoomMeetingId?: string | null;
  /** @nullable */
  zoomJoinUrl?: string | null;
  /** @nullable */
  zoomStartUrl?: string | null;
  status: MeetingStatus;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  rescheduledFrom?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  emailSentAt?: string | null;
  /** @nullable */
  reminder24SentAt?: string | null;
  /** @nullable */
  reminder1SentAt?: string | null;
  /** @nullable */
  cancellationEmailSentAt?: string | null;
  /** @nullable */
  rescheduleEmailSentAt?: string | null;
  createdAt: string;
}

export type MeetingInputMeetingType = typeof MeetingInputMeetingType[keyof typeof MeetingInputMeetingType];


export const MeetingInputMeetingType = {
  client: 'client',
  internal: 'internal',
  personal: 'personal',
} as const;

export interface MeetingInput {
  meetingType?: MeetingInputMeetingType;
  clientName: string;
  clientEmail: string;
  cc?: string[];
  bcc?: string[];
  meetingTitle: string;
  description?: string;
  meetingDate: string;
  meetingTime: string;
  duration?: number;
  timezone?: string;
  notes?: string;
}

export type MeetingUpdateMeetingType = typeof MeetingUpdateMeetingType[keyof typeof MeetingUpdateMeetingType];


export const MeetingUpdateMeetingType = {
  client: 'client',
  internal: 'internal',
  personal: 'personal',
} as const;

export type MeetingUpdateStatus = typeof MeetingUpdateStatus[keyof typeof MeetingUpdateStatus];


export const MeetingUpdateStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface MeetingUpdate {
  meetingType?: MeetingUpdateMeetingType;
  clientName?: string;
  clientEmail?: string;
  cc?: string[];
  bcc?: string[];
  meetingTitle?: string;
  description?: string;
  meetingDate?: string;
  meetingTime?: string;
  duration?: number;
  timezone?: string;
  notes?: string;
  status?: MeetingUpdateStatus;
}

export interface CancelMeetingInput {
  reason: string;
}

export interface RescheduleMeetingInput {
  meetingDate: string;
  meetingTime: string;
  reason?: string;
}

export type UpdateStatusInputStatus = typeof UpdateStatusInputStatus[keyof typeof UpdateStatusInputStatus];


export const UpdateStatusInputStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface UpdateStatusInput {
  status: UpdateStatusInputStatus;
}

export interface MeetingStats {
  total: number;
  upcoming: number;
  completed: number;
  cancelled: number;
  thisWeek: number;
  thisMonth: number;
}